# openapi/evermile.yaml
openapi: 3.0.2
info:
  contact:
    email: support@evermile.io
    name: API Support
    url: https://www.evermile.io/support
  description: This is Evermile's commercial API for handling delivery quotes and
    orders
  license:
    name: All rights reserved Evermile.io
    url: https://www.evermile.io/terms
  termsOfService: https://www.evermile.io/terms
  title: Evermile Commercial Quotes and Booking API
  version: "1.0"
servers:
- url: "https://api.{api-subdomain}.evermile.io/v1/commercial"
  variables:
    api-subdomain:
      default: prod
      description: The stage subdomain
      enum:
      - prod
      - sandbox
security:
- Sandbox:
  - read
  - order
- Prod:
  - read
  - order
paths:
...
  /webhook:
    get:
      description: Get the list of all registered webhooks
      parameters:
      - description: "The merchant ID, if using a client credentials token. Will be\
          \ ignored with a regular user token."
        in: header
        name: X-EVERMILE-MERCHANT-ID
        required: false
        schema:
          format: uuid
          type: string
      - description: A trace ID for tracing the request through the Evermile platform
        in: header
        name: X-EVERMILE-TRACE-ID
        required: false
        schema:
          type: string
      - description: The auth header specified in the URL registration request (the
          actual header name is up to the user)
        in: header
        name: X-EVERMILE-AUTH-HEADER
        required: true
        schema:
          type: string
      - description: The webhook topic this callback is referring to
        in: header
        name: X-EVERMILE-WEBHOOK-TOPIC
        required: true
        schema:
          $ref: '#/components/schemas/webhookTopic'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/webhookGetResponse'
          description: OK
        "401":
          content:
            application/json:
              schema:
                example: Unauthorized
                type: string
          description: Unauthorized access
        "409":
          content:
            application/json:
              schema:
                example: Too many requests
                type: string
          description: Rate limit exceeded
      summary: List all webhooks
      tags:
      - Webhooks
    parameters:
    - description: "The merchant ID, if using a client credentials token. Will be\
        \ ignored with a regular user token."
      in: header
      name: X-EVERMILE-MERCHANT-ID
      required: false
      schema:
        format: uuid
        type: string
    - description: A trace ID for tracing the request through the Evermile platform
      in: header
      name: X-EVERMILE-TRACE-ID
      required: false
      schema:
        type: string
    - description: The auth header specified in the URL registration request (the
        actual header name is up to the user)
      in: header
      name: X-EVERMILE-AUTH-HEADER
      required: true
      schema:
        type: string
    - description: The webhook topic this callback is referring to
      in: header
      name: X-EVERMILE-WEBHOOK-TOPIC
      required: true
      schema:
        $ref: '#/components/schemas/webhookTopic'
    post:
      callbacks:
        order.created:
          '{$request.body#/url}':
            post:
              description: Callback on order creation
              operationId: order_created_Post
              requestBody:
                content:
                  application/json:
                    schema:
                      $ref: '#/components/schemas/orderCreatedCallbackResp'
                description: The order creation update callback
              responses:
                "202":
                  description: |
                    Your server implementation should return this HTTP status code
                    if the data was received successfully
              summary: Created
              x-callback-request: true
        order.updated:
          '{$request.body#/url}':
            post:
              description: Callback on order update
              operationId: order_updated_Post
              requestBody:
                content:
                  application/json:
                    schema:
                      $ref: '#/components/schemas/orderCreatedCallbackResp'
                description: The order update callback
              responses:
                "202":
                  description: |
                    Your server implementation should return this HTTP status code
                    if the data was received successfully
              summary: Order Updated
              x-callback-request: true
        order.cancelled:
          '{$request.body#/url}':
            post:
              description: Callback on order cancellation
              operationId: order_cancelled_Post
              requestBody:
                content:
                  application/json:
                    schema:
                      $ref: '#/components/schemas/orderCancelledCallbackResp'
                description: The order cancellation callback
              responses:
                "202":
                  description: |
                    Your server implementation should return this HTTP status code
                    if the data was received successfully
              summary: Cancelled
              x-callback-request: true
        order.collected:
          '{$request.body#/url}':
            post:
              description: Callback on order collection by a delivery company
              operationId: order_collected_Post
              requestBody:
                content:
                  application/json:
                    schema:
                      $ref: '#/components/schemas/orderCollectedCallbackResp'
                description: The order collection callback
              responses:
                "202":
                  description: "Your server implementation should return this HTTP\
                    \ status code\nif the data was received successfully   \n"
              summary: Collected
              x-callback-request: true
        order.delivered:
          '{$request.body#/url}':
            post:
              description: Callback on successful order delivery by a delivery company
              operationId: order_delivered_Post
              requestBody:
                content:
                  application/json:
                    schema:
                      $ref: '#/components/schemas/orderDeliveredCallbackResp'
                description: The order delivered callback
              responses:
                "202":
                  description: |
                    Your server implementation should return this HTTP status code
                    if the data was received successfully
              summary: Delivered
              x-callback-request: true
        order.failed:
          '{$request.body#/url}':
            post:
              description: Callback on failed order delivery by a delivery company
              operationId: order_failed_Post
              requestBody:
                content:
                  application/json:
                    schema:
                      $ref: '#/components/schemas/orderFailedCallbackResp'
                description: The order failed callback
              responses:
                "202":
                  description: |
                    Your server implementation should return this HTTP status code
                    if the data was received successfully
              summary: Failed
              x-callback-request: true
        delivery.liveUpdate:
          '{$request.body#/url}':
            post:
              description: Callback on a live update of an ongoing delivery (when
                a courier is on the way to the end customer)
              operationId: delivery_liveUpdate_Post
              requestBody:
                content:
                  application/json:
                    schema:
                      $ref: '#/components/schemas/deliveryLiveUpdateCallbackResp'
                description: A live update callback body
              responses:
                "202":
                  description: |
                    Your server implementation should return this HTTP status code
                    if the data was received successfully
              summary: Live Update
              x-callback-request: true
      description: Register a webhook
      parameters:
      - description: "The merchant ID, if using a client credentials token. Will be\
          \ ignored with a regular user token."
        in: header
        name: X-EVERMILE-MERCHANT-ID
        required: false
        schema:
          format: uuid
          type: string
      - description: A trace ID for tracing the request through the Evermile platform
        in: header
        name: X-EVERMILE-TRACE-ID
        required: false
        schema:
          type: string
      - description: The auth header specified in the URL registration request (the
          actual header name is up to the user)
        in: header
        name: X-EVERMILE-AUTH-HEADER
        required: true
        schema:
          type: string
      - description: The webhook topic this callback is referring to
        in: header
        name: X-EVERMILE-WEBHOOK-TOPIC
        required: true
        schema:
          $ref: '#/components/schemas/webhookTopic'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/webhookRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/webhookPostResponse'
          description: Created
        "401":
          content:
            application/json:
              schema:
                example: Unauthorized
                type: string
          description: Unauthorized access
        "409":
          content:
            application/json:
              schema:
                example: Too many requests
                type: string
          description: Rate limit exceeded
      summary: Register webhook
      tags:
      - Webhooks
...
components:
...
  schemas:
...
    geoLocation:
      description: The geographical location (lat/lng)
      example:
        lng: -0.140666
        lat: 51.511221
      properties:
        lat:
          example: 51.511221
          format: double
          type: number
        lng:
          example: -0.140666
          format: double
          type: number
      required:
      - lat
      - lng
      type: object
...
    deliverySlot:
      description: A slot of time
      example:
        start: 2022-07-21T10:20:28Z
        end: 2022-07-21T17:32:28Z
      properties:
        start:
          description: The start time in ISO8601 format
          example: 2022-07-21T10:20:28Z
          format: date-time
          type: string
          x-oapi-codegen-extra-tags:
            csv: start
        end:
          description: The end time in ISO8601 format
          example: 2022-07-21T17:32:28Z
          format: date-time
          type: string
          x-oapi-codegen-extra-tags:
            csv: end
      required:
      - end
      - start
      type: object
...
    parcelType:
      default: Package
      description: |
        <u>The type of address</u>: <br>
        <table>
          <tr><td>Package</td>          <td>A packed box</td></tr>
          <tr><td>FlowerBouquet</td>    <td>A flower bouquet</td></tr>
        </table>
      enum:
      - Package
      - FlowerBouquet
      example: Package
      type: string
      x-oapi-codegen-extra-tags:
        csv: Type
...
    handlingInstruction:
      description: "<u>Handling instructions for an item</u>: <br>            \n<table>\n\
        \  <tr><td>alcohol</td>                <td>Parcel contains alcohol</td></tr>\n\
        \  <tr><td>cake</td>                   <td>Parcel contains a cake</td></tr>\n\
        \  <tr><td>flowers</td>                <td>Parcel contains flowers</td></tr>\n\
        \  <tr><td>fluids</td>                 <td>Parcel contains fluids (e.g. drinks)</td></tr>\n\
        \  <tr><td>fragile</td>                <td>Parcel contains fragile items</td></tr>\n\
        \  <tr><td>hot</td>                    <td>Parcel Contents are hot (e.g. food)</td></tr>\n\
        \  <tr><td>perishable</td>             <td>Parcel contains perishable items</td></tr>\n\
        \  <tr><td>prescriptionDrugs</td>      <td>Parcel contains prescription drugs</td></tr>\n\
        \  <tr><td>requiresRefrigeration</td>  <td>Parcel requires refrigeration</td></tr>\n\
        \  <tr><td>rightSideUp</td>            <td>Parcel must be kept right-side-up\
        \ (do not flip)</td></tr>\n  <tr><td>temperatureControlled</td>  <td>Parcel\
        \ has a temperature control device</td></tr>\n</table>\n"
      enum:
      - alcohol
      - cake
      - flowers
      - fluids
      - fragile
      - hot
      - perishable
      - prescriptionDrugs
      - requiresRefrigeration
      - rightSideUp
      - temperatureControlled
      type: string
    proofOfDeliveryRequirement:
      description: |
        An optional requirement for proof of delivery.<br>
        <b>NOTE:</b>Requiring a proof of delivery may limit optimization options and result in more expensive delivery quotes.<br>
        Will default to <b>none</b> if this parameter is not set.
        <table>
        <tr><td>signature</td>   <td>An electronic signature is required</td></tr>
        <tr><td>parcelPhoto</td> <td>A digital photo of the delivered parcel is required</td></tr>
        <tr><td>photoId</td>     <td>A digital photo of the recipient's photo ID is required</td></tr>
        <tr><td>any</td>         <td>Any proof of delivery is acceptable, but required</td></tr>
        <tr><td>none</td>        <td>No proof of delivery is required</td></tr>
        </table>
      enum:
      - signature
      - parcelPhoto
      - photoId
      - any
      - none
      type: string
...
    customer:
      description: Customer details
      example:
        phone: "447883196637"
        name: name
        email: johnsmith@email.com
      properties:
        name:
          description: The customer name
          type: string
        phone:
          description: The phone number of the contact person at address
          example: "447883196637"
          type: string
        email:
          description: The email address of the contact person at address
          example: johnsmith@email.com
          type: string
      required:
      - name
      type: object
...
    handoffType:
      enum:
      - none
      - receive
      - handoff
      example: receive
      type: string
...
    webhookTopic:
      description: |
        <u>The webhook topic</u>: <br>
        <table>
          <tr><td>order.created</td>        <td>A new order was created</td></tr>
          <tr><td>order.updated</td>        <td>An order was updated</td></tr>
          <tr><td>order.cancelled</td>      <td>An order was cancelled</td></tr>
          <tr><td>order.collected</td>      <td>An order was collected from the pickup location</td></tr>
          <tr><td>order.delivered</td>      <td>An order was delivered to the end customer</td></tr>
          <tr><td>order.failed</td>         <td>An order delivery failed</td></tr>
          <tr><td>delivery.liveUpdate</td>  <td>A live update of an ongoing delivery</td></tr>
        </table>
      enum:
      - order.created
      - order.updated
      - order.cancelled
      - order.collected
      - order.delivered
      - order.failed
      - delivery.liveUpdate
      example: order.delivered
      type: string
    webhookGetResponse:
      description: The response for a get all webhooks request
      example:
        webhooks:
        - authHeader: X-Auth-From-Evermile
          authHeaderKey: This is Me!
          topic: order.delivered
          id: f9bc2fa1-fa33-4b79-8fe5-f4fbe24c63ad
          url: https://my.shop.com/webhook
          lastSuccess: 2022-12-05T10:20:28Z
          status: enabled
          failuresSinceLastSuccess: 0
        - authHeader: X-Auth-From-Evermile
          authHeaderKey: This is Me!
          topic: order.delivered
          id: f9bc2fa1-fa33-4b79-8fe5-f4fbe24c63ad
          url: https://my.shop.com/webhook
          lastSuccess: 2022-12-05T10:20:28Z
          status: enabled
          failuresSinceLastSuccess: 0
      properties:
        webhooks:
          description: The list of webhooks
          items:
            $ref: '#/components/schemas/webhook'
          type: array
      type: object
    webhook:
      description: Details for a registered webhook
      example:
        authHeader: X-Auth-From-Evermile
        authHeaderKey: This is Me!
        topic: order.delivered
        id: f9bc2fa1-fa33-4b79-8fe5-f4fbe24c63ad
        url: https://my.shop.com/webhook
        lastSuccess: 2022-12-05T10:20:28Z
        status: enabled
        failuresSinceLastSuccess: 0
      properties:
        id:
          description: The Evermile ID for this webhook
          example: f9bc2fa1-fa33-4b79-8fe5-f4fbe24c63ad
          format: uuid
          type: string
        topic:
          $ref: '#/components/schemas/webhookTopic'
        url:
          description: The registered webhook URL.
          example: https://my.shop.com/webhook
          format: url
          type: string
        status:
          description: |
            <u>The webhook status</u>: <br>
            <table>
              <tr><td>enabled</td>     <td>The webhook is active</td></tr>
              <tr><td>disabled</td>    <td>The webhook is disabled (due to too many delivery failures</td></tr>
            </table>
          enum:
          - enabled
          - disabled
          type: string
        lastSuccess:
          description: "The last time the webhook call was successful, in ISO8601\
            \ format"
          example: 2022-12-05T10:20:28Z
          format: date-time
          type: string
        failuresSinceLastSuccess:
          description: The number of failures in calling the webhook since the last
            success
          example: 0
          format: int32
          minimum: 0
          type: integer
        authHeader:
          description: The header to use for sending a secret string authenticating
            a webhook call
          example: X-Auth-From-Evermile
          type: string
        authHeaderKey:
          description: The key to set in the header
          example: This is Me!
          type: string
      required:
      - authHeader
      - authHeaderKey
      - failuresSinceLastSuccess
      - id
      - status
      - topic
      - url
      type: object
    webhookRequest:
      description: A request for creating a webhook
      properties:
        topic:
          $ref: '#/components/schemas/webhookTopic'
        url:
          description: The URL to send the webhook to. Must be an **https** endpoint
          example: https://my.shop.com/webhook
          format: url
          type: string
        authHeader:
          description: A header to use for sending a secret string authenticating
            a webhook call
          example: X-Auth-From-Evermile
          type: string
        authHeaderKey:
          description: The key to set in the header
          example: This is Me!
          type: string
      required:
      - authHeader
      - authHeaderKey
      - topics
      - url
      type: object
    webhookPostResponse:
      description: The response from a webhook creation request
      example:
        webhook:
          authHeader: X-Auth-From-Evermile
          authHeaderKey: This is Me!
          topic: order.delivered
          id: f9bc2fa1-fa33-4b79-8fe5-f4fbe24c63ad
          url: https://my.shop.com/webhook
          lastSuccess: 2022-12-05T10:20:28Z
          status: enabled
          failuresSinceLastSuccess: 0
        secret: ""
      properties:
        webhook:
          $ref: '#/components/schemas/webhook'
        secret:
          description: The public key from a keypair used to validate the HMAC signature
            for a call from this webhook.
          example: ""
          type: string
      required:
      - secret
      - webhook
      type: object
    orderCreatedCallbackResp:
      properties:
        timestamp:
          description: The time this callback was sent in ISO8601 format
          example: 2022-12-05T10:22:28Z
          format: date-time
          type: string
        order:
          $ref: '#/components/schemas/orderDetails_1'
      required:
      - order
      - timestamp
      type: object
...
    orderDetails_1:
      description: The details of a booked order
      properties:
        id:
          description: The ID of the order
          example: f9bc2fa1-fa33-4b79-8fe5-f4fbe24c6edd
          format: uuid
          type: string
        externalOrderId:
          description: An external order ID attached to this order
          example: Order_1234
          type: string
        storeId:
          description: The ID of the e-commerce store for this order (if exists)
          example: f9bc2fa1-fa33-4b79-8fe5-f4fbe24c6edd
          format: uuid
          type: string
        storeName:
          description: The name of the store for this order (if exists)
          example: The Magic Wine Store
          type: string
        status:
          $ref: '#/components/schemas/orderExtendedStatus_1'
        communicatedPickup:
          $ref: '#/components/schemas/deliverySlot'
        communicatedDropoff:
          $ref: '#/components/schemas/deliverySlot'
        actualPickup:
          description: The actual pickup time in ISO8601 format (for packages picked
            up)
          example: 2022-07-21T10:25:28Z
          format: date-time
          type: string
        actualDropoff:
          description: The actual dropoff time in ISO8601 format (for packages dropped
            off)
          example: 2022-07-21T11:15:28Z
          format: date-time
          type: string
        pickupEta:
          description: The expected pickup time (if known)
          example: 2022-07-21T10:25:28Z
          format: date-time
          type: string
        dropoffEta:
          description: The expected dropoff time (if known)
          example: 2022-07-21T12:15:28Z
          format: date-time
          type: string
        courierLocation:
          $ref: '#/components/schemas/geoLocation'
        origin:
          $ref: '#/components/schemas/address_2'
        destination:
          $ref: '#/components/schemas/address_2'
        customer:
          $ref: '#/components/schemas/customer'
        deliveryCompany:
          description: The company that will perform the delivery (if known)
          example: Evermile
          type: string
        parcels:
          description: The list of parcels in this order
          items:
            $ref: '#/components/schemas/parcel_1'
          minItems: 1
          type: array
        value:
          $ref: '#/components/schemas/price_1'
        instructions:
          description: Additional handling instructions for the packages
          example: Fragile! Handle with care
          type: string
          x-oapi-codegen-extra-tags:
            csv: Instructions
        handling:
          description: Handling instructions for a parcel
          example:
          - alcohol
          - fluids
          - fragile
          items:
            $ref: '#/components/schemas/handlingInstruction'
          type: array
        proofOfDeliveryRequirement:
          default:
          - none
          description: Proof of delivery requirements for a parcel
          example:
          - signature
          - photoId
          items:
            $ref: '#/components/schemas/proofOfDeliveryRequirement'
          type: array
        trackingUrl:
          description: The carrier tracking URL for this order (if exists)
          example: https://tracking.evermile.io/order/f9bc2fa1-fa33-4b79-8fe5-f4fbe24c6edd
          type: string
        customerTrackingUrl:
          description: The Evermile customer tracking URL for this order (if exists)
          example: https://evml.io/FDS34RE2
          type: string
        orderTrackingInfo:
          $ref: '#/components/schemas/orderTrackingInfo_1'
        handoffType:
          $ref: '#/components/schemas/handoffType'
        handoffInfo:
          $ref: '#/components/schemas/handoffInfo_1'
        usedCredits:
          description: Whether credits were used to pay for this order
          example: true
          type: boolean
        proofOfDelivery:
          description: Proof of delivery for an order
          items:
            $ref: '#/components/schemas/proofOfDelivery'
          type: array
        notes:
          description: Courier notes from each delivery in this order
          items:
            type: string
          type: array
        merchantName:
          description: The name of the merchant for this order
          example: The Magic Wine Store
          type: string
        labelRequired:
          description: Whether label is required
          example: true
          type: boolean
      required:
      - communicatedDropoff
      - communicatedPickup
      - currency
      - customer
      - destination
      - externalOrderId
      - handoffType
      - id
      - labelRequired
      - merchantName
      - origin
      - parcels
      - status
      - usedCredits
      - value
      type: object
    orderExtendedStatus_1:
      description: An order status with reporting information
      properties:
        orderStatus:
          $ref: '#/components/schemas/orderStatus'
        reporter:
          description: Who reported this status
          example: John
          type: string
        reporting_time:
          description: The reporting time in ISO8601 format
          example: 2022-07-21T10:20:28Z
          format: date-time
          type: string
        manuallyUpdated:
          description: True if the status was manually changed by the reporter
          example: true
          type: boolean
      required:
      - manuallyUpdated
      - orderStatus
      - reporter
      - reporting_time
      type: object
    address_2:
      description: An address for picking up or dropping off
      properties:
        addressLine1:
          description: Street and house number for the address
          example: 32 Saville Row
          type: string
          x-oapi-codegen-extra-tags:
            csv: Address Line 1
        addressLine2:
          description: Additional address details such as flat number
          example: "12"
          type: string
          x-oapi-codegen-extra-tags:
            csv: Address Line 2
        city:
          description: The city for the address
          example: London
          type: string
          x-oapi-codegen-extra-tags:
            csv: City
        postalCode:
          description: The postal code for the address
          example: W1S 3PT
          type: string
          x-oapi-codegen-extra-tags:
            csv: Postal Code
        geoLocation:
          $ref: '#/components/schemas/geoLocation'
        type:
          $ref: '#/components/schemas/addressType'
      required:
      - addressLine1
      - city
      - postalCode
      - type
      type: object
    parcel_1:
      description: Data about the parcel
      properties:
        id:
          description: The id of the parcel object in evermile db
          format: uuid
          type: string
        dimensions:
          $ref: '#/components/schemas/dimensions'
        weightKg:
          description: Weight of the package in Kilograms
          example: 2.3
          format: float
          type: number
          x-oapi-codegen-extra-tags:
            csv: Weight (kg)
        parcelType:
          $ref: '#/components/schemas/parcelType'
        itemsList:
          items:
            $ref: '#/components/schemas/item_1'
          minItems: 1
          type: array
      required:
      - dimensions
      - parcelType
      - weightKg
      type: object
...
    price_1:
      description: A price object
      properties:
        value:
          description: The item's value in cents/pence. E.G. "500" = 5 GBP
          example: 500
          format: int64
          type: integer
        currency:
          example: GBP
          type: string
        discount:
          $ref: '#/components/schemas/discount'
      required:
      - currency
      - value
      type: object
...
    handoffInfo_1:
      description: Information about the handoff
      properties:
        contactInfo:
          $ref: '#/components/schemas/handoffContactInfo_1'
        handoffTime:
          $ref: '#/components/schemas/optionalTimeWindow'
        walkingDistanceMeters:
          description: walking distance in meters
          example: 100
          format: int64
          type: integer
      required:
      - contactInfo
      - handoffTime
      - walkingDistanceMeters
      type: object
...
    orderCancelledCallbackResp:
      properties:
        timestamp:
          description: The time this callback was sent in ISO8601 format
          example: 2022-12-05T10:22:28Z
          format: date-time
          type: string
        orderId:
          description: The ID of the order
          example: f9bc2fa1-fa33-4b79-8fe5-f4fbe24c6edd
          format: uuid
          type: string
        cancellationFee:
          $ref: '#/components/schemas/price_1'
      required:
      - orderId
      - timestamp
      type: object
    orderCollectedCallbackResp:
      properties:
        timestamp:
          description: The time this callback was sent in ISO8601 format
          example: 2022-12-05T10:22:28Z
          format: date-time
          type: string
        orderId:
          description: The ID of the order
          example: f9bc2fa1-fa33-4b79-8fe5-f4fbe24c6edd
          format: uuid
          type: string
      required:
      - orderId
      - timestamp
      type: object
    orderDeliveredCallbackResp:
      properties:
        timestamp:
          description: The time this callback was sent in ISO8601 format
          example: 2022-12-05T10:22:28Z
          format: date-time
          type: string
        orderId:
          description: The ID of the order
          example: f9bc2fa1-fa33-4b79-8fe5-f4fbe24c6edd
          format: uuid
          type: string
      required:
      - orderId
      - timestamp
      type: object
    orderFailedCallbackResp:
      properties:
        timestamp:
          description: The time this callback was sent in ISO8601 format
          example: 2022-12-05T10:22:28Z
          format: date-time
          type: string
        orderId:
          description: The ID of the order
          example: f9bc2fa1-fa33-4b79-8fe5-f4fbe24c6edd
          format: uuid
          type: string
        failureReason:
          description: An optional string detailing the failure reason
          example: Customer could not be reached
          type: string
      required:
      - orderId
      - timestamp
      type: object
    deliveryLiveUpdateCallbackResp:
      properties:
        timestamp:
          description: The time this callback was sent in ISO8601 format
          example: 2022-12-05T10:22:28Z
          format: date-time
          type: string
        orderId:
          description: The ID of the order
          example: f9bc2fa1-fa33-4b79-8fe5-f4fbe24c6edd
          format: uuid
          type: string
        eta:
          description: The expected delivery time
          example: 2022-12-06T17:32:28Z
          format: date-time
          type: string
        courierLocation:
          $ref: '#/components/schemas/geoLocation'
      required:
      - orderId
      - timestamp
      type: object
...
  securitySchemes:
    Sandbox:
      description: Sandbox OAuth2 client credentials flow
      flows:
        clientCredentials:
          scopes:
            order: Order deliveries
            read: Get quotes and read information about deliveries
          tokenUrl: https://auth.sandbox.evermile.io/oauth2/token
      type: oauth2
    Prod:
      description: Production OAuth2 client credentials flow
      flows:
        clientCredentials:
          scopes:
            order: Order deliveries
            read: Get quotes and read information about deliveries
          tokenUrl: https://auth.prod.evermile.io/oauth2/token
      type: oauth2